arm: Support for Base Priority Mask Register
Currently there is no emulation of the ARM BASEPRI register. Only reading/writing
the register is implemented, but no functionality.
The BASEPRI register defines the minimum priority for exception processing.
When BASEPRI is set to a nonzero value, it prevents the activation of all
exceptions with the same or lower priority level as the BASEPRI value.

This patchset implements the expected functionality. BASEPRI reads/writes are
implemented via helper routines (MSR/MRS), so a write to BASEPRI will result
in a call gic_update with a new BASEPRI value. Only writes to BASEPRI are tracked,
and gic_update is called only if BASEPRI has changed.

http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.dui0552a/CHDBIBGJ.html

Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>
Signed-off-by: Anas Nashif <anas.nashif@intel.com>