qemu: Determinism fixes
When sources are included within debug information, a couple of areas of the
qemu build are not reproducible due to either full buildpaths or timestamps.

Replace the full paths with relative ones. I couldn't figure out how to get
meson to pass relative paths but we can fix that in the script.

Upstream-Status: Pending [some version of all/part of this may be accepted]
RP 2021/3/1